Add activity ok

+++ database/migrations/2020_09_12_184858_create_activities_table.php
@@ -17,7 +17,7 @@ class CreateActivitiesTable extends Migration
             $table->dateTime('started_at')->nullable()->comment('Activity start time');
             $table->boolean('shuffle_questions')->default(false);
             $table->boolean('shuffle_propositions')->default(false);
-            $table->integer('seed')->default(0)->comment('For random generator');
+            $table->bigInteger('seed')->default(0)->comment('For random generator');
             $table->boolean('hidden')->default(false);
 
             $table->timestamps();
